Moved obsolete quenching modules into obsolete
