+++ diffmah/obsolete/tests/test_quenched_fraction.py
